+++ src/include/dbgcfg.h
@@ -6,10 +6,10 @@
 // and a couple of macros to do stuff when the system is build with DEBUG=1 (or
 // 2). Set the DEBUG flag in the Makefile :)
 #if defined(DEBUG)
-void dumpVal(unsigned long v, const char* header = nullptr);
+void dumpVal(uint32_t v, const char* header = nullptr);
 void dumpHex(uint8_t v, const char* header = nullptr);
 void dumpHex(uint16_t v, const char* header = nullptr);
-void dumpHex(unsigned long v, const char* header = nullptr);
+void dumpHex(uint32_t v, const char* header = nullptr);
 void dumpHex(uint64_t v, const char* header = nullptr);
 void dumpHex(bool b, const char* header = nullptr);
 #define DBG(a) a
